Create Intersection of two array
Problem Statement
Given an integer array A of size N which is sorted. You are also given another integer array B of size M which may or may not be sorted but all the elements of B are distinct. Intersection of two arrays A and B is a sorted array C which contains only distinct elements and all elements of C are present in both A and B. Find the intersection of A and B.
Note: There is atleast one common element in A and B.
Input
First line contains an integer N.
Next line contains N space separated integers denoting elements of array A.
Next line contains an integer M.
Next line contains M space separated integers denoting elements of array B.

Constraints
1 <= N <= 10^5
0 <= Ai <= 10^5
1 <= M <= 10^5
0 <= Bi <= 10^5
Output
Print K space separated integers denoting the elements of array C - intersection of arrays A and B.
Example
Sample Input 1:
3
1 2 3
3
2 1 3

Output
1 2 3

Explanation:
1,2 and 3 are present in both arrays.